+++ src/components/modality-section/modality-options.css
