+++ src/main.jsx
@@ -1,8 +1,7 @@
 import App from './App.jsx'
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import ModalityOptions from './components/modality-options/modality-options.jsx'
-import ModalitySection from './components/modality-options/ModalitySection.jsx'
+import ModalitySection from './components/modality-section/modality-section'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
